requirements.txt
Django==1.4
django-social-auth
django-tastypie
slumber
